Shift responsibility of IDFA collection to clients
This is a backport from segmentio/analytics-ios (https://github.com/segmentio/analytics-ios/pull/892).

Any references to Apple's AdSupport framework might trip the App Store's static analysis, making it a bit unclear if SDK users can confidently say that their apps do not use the IDFA. By moving the responsibility to collect the IDFA out of the PostHog SDK and into the client app, the AdSupport references are gone and the question is easily answered.